requirements.txt
click==6.7
Flask==1.0.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.3
pymongo==3.7.1
python-dotenv==0.9.1
six==1.11.0
Werkzeug==0.14.1
WTForms==2.2.1
